improve godoc formatting
