+++ pkg/collections/wrappers.go
@@ -2,11 +2,11 @@
 // Use of this source code is governed by the MIT
 // license that can be found in the LICENSE file.
 
-// wrappers.go contains collection methods that wrap existing functions
-// provided by the go slices library to provide syntatic sugar and enable
-// function chaining, for example:
+// wrappers.go contains collection methods that wrap functions
+// provided by the Go slices library. This file is just synthatic sugar to
+// enable function chaining i.e.:
 //
-// collection.Reverse().Concat(c2).Sort()
+//    collection.Reverse().Concat(c2).Sort()
 
 package collections
 
@@ -15,10 +15,12 @@ import (
 	"slices"
 )
 
+// All returns an interator over all (index,value) pairs of the underlying slice.
 func (c *Collection[T]) All() iter.Seq2[int, T] {
 	return slices.All(c.elements)
 }
 
+// Values returns an iterator over all values of the underlying slice.
 func (c *Collection[T]) Values() iter.Seq[T] {
 	return slices.Values(c.elements)
 }
